CIV-10144 Refactor Judge Make order docmosis generator
