+++ src/main/java/uk/gov/hmcts/reform/civil/service/docmosis/directionorder/DirectionOrderGenerator.java
@@ -2,32 +2,26 @@ package uk.gov.hmcts.reform.civil.service.docmosis.directionorder;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
-import org.apache.commons.lang.StringUtils;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.civil.enums.YesOrNo;
-import uk.gov.hmcts.reform.civil.enums.dq.GAByCourtsInitiativeGAspec;
 import uk.gov.hmcts.reform.civil.model.CaseData;
 import uk.gov.hmcts.reform.civil.model.docmosis.DocmosisDocument;
 import uk.gov.hmcts.reform.civil.model.docmosis.judgedecisionpdfdocument.JudgeDecisionPdfDocument;
 import uk.gov.hmcts.reform.civil.model.documents.CaseDocument;
 import uk.gov.hmcts.reform.civil.model.documents.DocumentType;
 import uk.gov.hmcts.reform.civil.model.documents.PDF;
+import uk.gov.hmcts.reform.civil.service.docmosis.DocmosisService;
 import uk.gov.hmcts.reform.civil.service.docmosis.DocmosisTemplates;
 import uk.gov.hmcts.reform.civil.service.docmosis.DocumentGeneratorService;
 import uk.gov.hmcts.reform.civil.service.docmosis.ListGeneratorService;
 import uk.gov.hmcts.reform.civil.service.docmosis.TemplateDataGenerator;
 import uk.gov.hmcts.reform.civil.service.documentmanagement.DocumentManagementService;
-import uk.gov.hmcts.reform.idam.client.IdamClient;
-import uk.gov.hmcts.reform.idam.client.models.UserDetails;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Date;
 import java.util.Map;
-import java.util.Objects;
 
 import static uk.gov.hmcts.reform.civil.service.docmosis.DocmosisTemplates.DIRECTION_ORDER;
-import static uk.gov.hmcts.reform.civil.service.docmosis.DocumentGeneratorService.DATE_FORMATTER;
 import static uk.gov.hmcts.reform.civil.service.docmosis.generalorder.GeneralOrderGenerator.showRecital;
 import static uk.gov.hmcts.reform.civil.utils.DateFormatterUtil.getFormattedDate;
 
@@ -38,17 +32,16 @@ public class DirectionOrderGenerator implements TemplateDataGenerator<JudgeDecis
     private final DocumentManagementService documentManagementService;
     private final DocumentGeneratorService documentGeneratorService;
     private final ListGeneratorService listGeneratorService;
-    private final IdamClient idamClient;
     private final ObjectMapper mapper;
 
+    private final DocmosisService docmosisService;
+
     public CaseDocument generate(CaseData caseData, String authorisation) {
-        UserDetails userDetails = idamClient.getUserDetails(authorisation);
-        String judgeNameTitle = userDetails.getFullName();
 
         JudgeDecisionPdfDocument templateData = getTemplateData(caseData);
 
         Map<String, Object> map = templateData.toMap(mapper);
-        map.put("judgeNameTitle", judgeNameTitle);
+        map.put("judgeNameTitle", docmosisService.getJudgeNameTitle(authorisation));
         templateData = mapper.convertValue(map, JudgeDecisionPdfDocument.class);
 
         DocmosisTemplates docmosisTemplate = getDocmosisTemplate();
@@ -87,48 +80,11 @@ public class DirectionOrderGenerator implements TemplateDataGenerator<JudgeDecis
                 .judgeDirection(caseData.getJudicialDecisionMakeOrder().getDirectionsText())
                 .reasonForDecision(caseData.getJudicialDecisionMakeOrder().getReasonForDecisionText())
                 .submittedOn(getFormattedDate(new Date()))
-                .reasonAvailable(reasonAvailable(caseData))
-                .reasonForDecision(populateJudgeReasonForDirection(caseData))
-                .judicialByCourtsInitiative(populateJudicialByCourtsInitiative(caseData));
+                .reasonAvailable(docmosisService.reasonAvailable(caseData))
+                .reasonForDecision(docmosisService.populateJudgeReason(caseData))
+                .judicialByCourtsInitiative(docmosisService.populateJudicialByCourtsInitiative(caseData));
 
         return judgeDecisionPdfDocumentBuilder.build();
-    }
-
-    private YesOrNo reasonAvailable(CaseData caseData) {
-        if (Objects.nonNull(caseData.getJudicialDecisionMakeOrder().getShowReasonForDecision())
-            && caseData.getJudicialDecisionMakeOrder().getShowReasonForDecision().equals(YesOrNo.NO)) {
-            return YesOrNo.NO;
-        }
-        return YesOrNo.YES;
-    }
-
-    private String populateJudgeReasonForDirection(CaseData caseData) {
-        if (Objects.nonNull(caseData.getJudicialDecisionMakeOrder().getShowReasonForDecision())
-            && caseData.getJudicialDecisionMakeOrder().getShowReasonForDecision().equals(YesOrNo.NO)) {
-            return "";
-        }
-        return caseData.getJudicialDecisionMakeOrder().getReasonForDecisionText() != null
-            ? caseData.getJudicialDecisionMakeOrder().getReasonForDecisionText()
-            : "";
-    }
-
-    public String populateJudicialByCourtsInitiative(CaseData caseData) {
-
-        if (caseData.getJudicialDecisionMakeOrder().getJudicialByCourtsInitiative().equals(GAByCourtsInitiativeGAspec
-                                                                                               .OPTION_3)) {
-            return StringUtils.EMPTY;
-        }
-
-        if (caseData.getJudicialDecisionMakeOrder().getJudicialByCourtsInitiative()
-            .equals(GAByCourtsInitiativeGAspec.OPTION_1)) {
-            return caseData.getJudicialDecisionMakeOrder().getOrderCourtOwnInitiative() + " "
-                .concat(caseData.getJudicialDecisionMakeOrder().getOrderCourtOwnInitiativeDate()
-                            .format(DATE_FORMATTER));
-        } else {
-            return caseData.getJudicialDecisionMakeOrder().getOrderWithoutNotice() + " "
-                .concat(caseData.getJudicialDecisionMakeOrder().getOrderWithoutNoticeDate()
-                            .format(DATE_FORMATTER));
-        }
     }
 
     private DocmosisTemplates getDocmosisTemplate() {
