+++ src/main/java/uk/gov/hmcts/reform/civil/service/docmosis/dismissalorder/DismissalOrderGenerator.java
@@ -2,10 +2,7 @@ package uk.gov.hmcts.reform.civil.service.docmosis.dismissalorder;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
-import org.apache.commons.lang.StringUtils;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.civil.enums.YesOrNo;
-import uk.gov.hmcts.reform.civil.enums.dq.GAByCourtsInitiativeGAspec;
 import uk.gov.hmcts.reform.civil.model.CaseData;
 import uk.gov.hmcts.reform.civil.model.docmosis.DocmosisDocument;
 import uk.gov.hmcts.reform.civil.model.docmosis.judgedecisionpdfdocument.JudgeDecisionPdfDocument;
@@ -23,10 +20,8 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Date;
 import java.util.Map;
-import java.util.Objects;
 
 import static uk.gov.hmcts.reform.civil.service.docmosis.DocmosisTemplates.DISMISSAL_ORDER;
-import static uk.gov.hmcts.reform.civil.service.docmosis.DocumentGeneratorService.DATE_FORMATTER;
 import static uk.gov.hmcts.reform.civil.utils.DateFormatterUtil.getFormattedDate;
 
 @Service
@@ -81,39 +76,10 @@ public class DismissalOrderGenerator implements TemplateDataGenerator<JudgeDecis
                 .dismissalOrder(caseData.getJudicialDecisionMakeOrder().getDismissalOrderText())
                 .submittedOn(getFormattedDate(new Date()))
                 .reasonAvailable(docmosisService.reasonAvailable(caseData))
-                .reasonForDecision(populateJudgeReasonForDismissal(caseData))
-                .judicialByCourtsInitiative(populateJudicialByCourtsInitiative(caseData));
+                .reasonForDecision(docmosisService.populateJudgeReason(caseData))
+                .judicialByCourtsInitiative(docmosisService.populateJudicialByCourtsInitiative(caseData));
 
         return judgeDecisionPdfDocumentBuilder.build();
-    }
-
-    private String populateJudgeReasonForDismissal(CaseData caseData) {
-        if (Objects.nonNull(caseData.getJudicialDecisionMakeOrder().getShowReasonForDecision())
-            && caseData.getJudicialDecisionMakeOrder().getShowReasonForDecision().equals(YesOrNo.NO)) {
-            return "";
-        }
-        return caseData.getJudicialDecisionMakeOrder().getReasonForDecisionText() != null
-            ? caseData.getJudicialDecisionMakeOrder().getReasonForDecisionText()
-            : "";
-    }
-
-    private String populateJudicialByCourtsInitiative(CaseData caseData) {
-
-        if (caseData.getJudicialDecisionMakeOrder().getJudicialByCourtsInitiative().equals(GAByCourtsInitiativeGAspec
-                                                                                               .OPTION_3)) {
-            return StringUtils.EMPTY;
-        }
-
-        if (caseData.getJudicialDecisionMakeOrder().getJudicialByCourtsInitiative()
-            .equals(GAByCourtsInitiativeGAspec.OPTION_1)) {
-            return caseData.getJudicialDecisionMakeOrder().getOrderCourtOwnInitiative() + " "
-                .concat(caseData.getJudicialDecisionMakeOrder().getOrderCourtOwnInitiativeDate()
-                            .format(DATE_FORMATTER));
-        } else {
-            return caseData.getJudicialDecisionMakeOrder().getOrderWithoutNotice() + " "
-                .concat(caseData.getJudicialDecisionMakeOrder().getOrderWithoutNoticeDate()
-                            .format(DATE_FORMATTER));
-        }
     }
 
     private DocmosisTemplates getDocmosisTemplate() {
